+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add ability to clear the boolean gate only (useful for debugging)
+* Added `FunWithFlags.get_flag/1`, to retrieve a flag struct. Useful for debugging.
 
 ## v0.10.1